requirements.txt
Django==1.8.5
django-autocomplete-light==2.2.10
django-crispy-forms==1.5.2
django-registration-redux==1.2
gunicorn==19.3.0
psycopg2==2.6.1
whitenoise==2.0.6
dj-database-url==0.3.0
